requirements-backend.txt
flask==2.3.3
pandas==2.1.0
yfinance==0.2.32
google-generativeai==0.3.1
